# API Layer/swagger_server/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: Smart App Group C - API Layer
  description: |-
    This is the API used for internal communication between services and by external applications and tools to allow data access.
    Made by the integration team: Andrea Alberti, Anna Francesca Montagnoli, Matteo Simone
  version: 1.1.1
servers:
- url: /
...
paths:
...
  /data/raw:
    get:
      tags:
      - data
      summary: "Gets raw sensors data filtered by machine ID, data type (as KPI ids),\
        \ and time window"
      description: "Access to basic KPI related to sensor data. The query can be restricted\
        \ to specific machines and/or KPIs given their ids and a time window. The\
        \ time window is interpreted as [from, to). Since KPIs are aggregated using\
        \ multiple functions, 'aggregation selector' determines the type of aggregation\
        \ to retrieve."
      operationId: data_get
      parameters:
      - name: machines
        in: header
        description: Specific machines to restrict the query to
        required: false
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/MachineIDFilter'
      - name: dataTypes
        in: header
        description: Specific basic kpi to restrict the query to
        required: false
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/DataTypeFilter'
      - name: aggregationSelector
        in: header
        description: Aggregation type to retrieve
        required: true
        style: simple
        explode: false
        schema:
          type: string
          example: sum
          enum:
          - sum
          - avg
          - min
          - max
      - name: from
        in: header
        description: The first timestamp of the time window of results
        required: false
        style: simple
        explode: false
        schema:
          type: string
          format: date-time
      - name: to
        in: header
        description: The last timestamp of the time window of results
        required: false
        style: simple
        explode: false
        schema:
          type: string
          format: date-time
      responses:
        "200":
          description: Succesful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SensorData'
                x-content-type: application/json
        "400":
          description: Invalid parameters
          content:
            application/json:
              schema:
                type: string
                example: Invalid parameters
                x-content-type: application/json
        "401":
          description: Access token not valid
          content:
            application/json:
              schema:
                type: string
                example: Unauthorized
                x-content-type: application/json
        "500":
          description: Can't connect to the requested service
          content:
            application/json:
              schema:
                type: string
                example: Connection Error
                x-content-type: application/json
      security:
      - api_token_auth: []
      x-openapi-router-controller: swagger_server.controllers.data_controller
  /data/preprocessed:
    get:
      tags:
      - data
      summary: "Gets preprocessed sensors data filtered by machine ID, data type (as\
        \ KPI ids), and time window"
      description: "Access to basic KPI related to preprocessed sensor data. The query\
        \ can be restricted to specific machines and/or KPIs given their ids and a\
        \ time window. The time window is interpreted as [from, to). Since KPIs are\
        \ aggregated using multiple functions, 'aggregation selector' determines the\
        \ type of aggregation to retrieve."
      operationId: data_preprocessed_get
      parameters:
      - name: machines
        in: header
        description: Specific machines to restrict the query to
        required: false
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/MachineIDFilter'
      - name: dataTypes
        in: header
        description: Specific basic kpi to restrict the query to
        required: false
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/DataTypeFilter'
      - name: aggregationSelector
        in: header
        description: Aggregation type to retrieve
        required: true
        style: simple
        explode: false
        schema:
          type: string
          example: sum
          enum:
          - sum
          - avg
          - min
          - max
      - name: from
        in: header
        description: The first timestamp of the time window of results
        required: false
        style: simple
        explode: false
        schema:
          type: string
          format: date-time
      - name: to
        in: header
        description: The last timestamp of the time window of results
        required: false
        style: simple
        explode: false
        schema:
          type: string
          format: date-time
      responses:
        "200":
          description: Succesful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SensorData'
                x-content-type: application/json
        "400":
          description: Invalid parameters
          content:
            application/json:
              schema:
                type: string
                example: Invalid parameters
                x-content-type: application/json
        "401":
          description: Access token not valid
          content:
            application/json:
              schema:
                type: string
                example: Unauthorized
                x-content-type: application/json
        "500":
          description: Can't connect to the requested service
          content:
            application/json:
              schema:
                type: string
                example: Connection Error
                x-content-type: application/json
      security:
      - api_token_auth: []
      x-openapi-router-controller: swagger_server.controllers.data_controller
  /data/{KPIID}/{machineID}:
    post:
      tags:
      - data
      summary: Adds new sensor data to the system
      description: Adds new aggregated values of a KPI for a given machine id. AggregationType
        and values must be of the same length. 400 error will be returned if the machine
        can't generate such KPI.
      operationId: data_post
      parameters:
      - name: KPIID
        in: path
        description: ID of the KPI
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/KPIID'
      - name: machineID
        in: path
        description: ID of the machine
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          example: 1
      - name: aggregationType
        in: header
        description: The aggregation types of the KPI values to be added
        required: true
        style: simple
        explode: false
        schema:
          type: array
          example:
          - sum
          - min
          items:
            type: string
            enum:
            - max
            - sum
            - min
            - avg
      - name: values
        in: header
        description: Aggregated values of the KPI
        required: true
        style: simple
        explode: false
        schema:
          type: array
          example:
          - 4.5
          - 1
          items:
            type: number
      responses:
        "200":
          description: Success Message
          content:
            application/json:
              schema:
                type: string
                example: Success
                x-content-type: application/json
        "400":
          description: Invalid parameters
          content:
            application/json:
              schema:
                type: string
                example: Invalid parameters
                x-content-type: application/json
        "401":
          description: Access token not valid
          content:
            application/json:
              schema:
                type: string
                example: Unauthorized
                x-content-type: application/json
        "404":
          description: KPI/machine not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_404'
        "500":
          description: Can't connect to the requested service
          content:
            application/json:
              schema:
                type: string
                example: Connection Error
                x-content-type: application/json
      security:
      - api_token_auth: []
      x-openapi-router-controller: swagger_server.controllers.data_controller
...
components:
  schemas:
...
    MachineNotFoundError:
      type: string
      description: Machine not found error message
      example: Machine not found
    KPINotFoundError:
      type: string
      description: KPI not found error message
      example: KPI not found
...
    ID:
      type: integer
      description: Numerical unique identifier of the resource
      format: int64
      example: 1
    KPIID:
      type: string
      description: A unique identifier of a KPI
      example: good_cycles
...
    DataTypeFilter:
      type: array
      description: List of KPI ids
      example:
      - power
      - good_cycles
      items:
        $ref: '#/components/schemas/KPIID'
    MachineIDFilter:
      type: array
      description: List of machine identifiers
      example:
      - 1
      - 2
      items:
        $ref: '#/components/schemas/ID'
...
    SensorData:
      type: object
      properties:
        machineID:
          $ref: '#/components/schemas/ID'
        kpi:
          $ref: '#/components/schemas/KPIID'
        timestampSeries:
          type: array
          items:
            type: string
            format: date-time
        valueSeries:
          type: array
          items:
            type: number
            example: 4.5
      description: Series of values of a KPI generated by a machine. The i-th timestamp
        is related to the i-th value.
      example:
        machineID: 1
        kpi: good_cycles
        valueSeries:
        - 4.5
        - 4.5
        timestampSeries:
        - 2000-01-23T04:56:07.000+00:00
        - 2000-01-23T04:56:07.000+00:00
...
    inline_response_404:
      oneOf:
      - $ref: '#/components/schemas/MachineNotFoundError'
      - $ref: '#/components/schemas/KPINotFoundError'
...
  securitySchemes:
    api_token_auth:
      type: http
      scheme: bearer
      x-bearerInfoFunc: swagger_server.controllers.authorization_controller.check_api_token_auth
    machine_auth:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: https://SA2024C.com/oauth/token
          scopes:
            read:machine: Read machine data
            write:machine: Write machine data
            read:KPI: Read KPI data
            write:KPI: Write KPI data
            read:user: Access user data
            write:user: Modify user data
            read:alert: Read alert data
            write:alert: Edit alert data
            read:report: Read SMO report data
            write:report: Edit report data
            read:aiquery: Get ai query responses
            write:aiquery: Request new ai queries
            read:data: Access raw data
        authorizationCode:
          authorizationUrl: https://SA2024C.com/oauth/authorize
          tokenUrl: https://SA2024C.com/oauth/token
          scopes:
            read:machine: Read machine data
            write:machine: Write machine data
            read:KPI: Read KPI data
            write:KPI: Write KPI data
            read:user: Access user data
            write:user: Modify user data
            read:alert: Read alert data
            write:alert: Edit alert data
            read:report: Read SMO report data
            write:report: Edit report data
            read:data: Access raw data
      x-tokenInfoFunc: swagger_server.controllers.authorization_controller.check_machine_auth
      x-scopeValidateFunc: swagger_server.controllers.authorization_controller.validate_scope_machine_auth